+++ mf-vimeo-video.php
@@ -4,7 +4,7 @@ Plugin Name: Vimeo Video CPT
 Plugin URI: https://github.com/miszou/vimeo-video
 GitHub Plugin URI: miszou/vimeo-video
 Description: Registers a Vimeo Video custom post type with Media Tag taxonomy, filterable and searchable via REST API.
-Version: 0.4.7
+Version: 0.4.8
 Author: miszou
 Text Domain: mf-vimeo-video
 */
@@ -84,6 +84,7 @@ function mfvv_register_video_cpt()
             "excerpt",
             "author",
             "custom-fields",
+            "page-attributes",
         ],
         "rewrite" => ["slug" => "videos"],
         "show_in_rest" => true,
@@ -186,7 +187,7 @@ function mfvv_admin_enqueue($hook)
         "mfvv-admin",
         plugins_url("assets/js/admin.js", __FILE__),
         ["wp-data"],
-        "0.4.7",
+        "0.4.8",
         true,
     );
     wp_localize_script("mfvv-admin", "mfvvAdmin", [